+++ spec/luna_park/errors/adpative_spec.rb
@@ -106,38 +106,97 @@ module LunaPark
         end
       end
 
-      context 'message is defined on initialization error object' do
-        let(:msg) { 'defined message' }
-        subject(:message) { described_class.new(msg).message }
+      context 'message is given on initialization' do
+        subject(:message) { described_class.new('given message').message }
 
         it 'should be eq class name' do
-          is_expected.to eq msg
+          is_expected.to eq 'given message'
         end
       end
 
-      context 'message defined in class' do
-        let(:error) do
+      context 'message is defined in class by string' do
+        let(:error_class) do
           Class.new(described_class) do
-            message 'Class message'
+            message 'Static default message'
           end
         end
 
         context 'message is not defined on initialization error object' do
-          subject(:message) { error.new.message }
+          subject(:message) { error_class.new.message }
 
           it 'should be eq class name' do
-            is_expected.to eq 'Class message'
+            is_expected.to eq 'Static default message'
+          end
+        end
+      end
+
+      context 'message defined in class by block' do
+        subject(:message) { error_class.new(foo: 'FOO').message }
+
+        let(:error_class) do
+          Class.new(described_class) do
+            message { |d| "Dynamic message with variable '#{d[:foo]}'" }
           end
         end
 
-        context 'message is defined on initialization error object' do
-          let(:msg) { 'defined message' }
-          subject(:message) { error.new(msg).message }
+        it 'should be expected dynamic message' do
+          is_expected.to eq "Dynamic message with variable 'FOO'"
+        end
+      end
 
-          it 'should be eq class name' do
-            is_expected.to eq msg
+      context 'message is defined with i18n' do
+        let(:error) { error_class.new(foo: 'FOO') }
+        let(:error_class) do
+          Class.new(described_class) do
+            message 'Static default message', i18n_key: 'errors.example'
           end
         end
+
+        it 'is equal to expected message at default locale' do
+          expect(error.message).to eq 'Example'
+        end
+
+        it 'is equal to expected message at given locale' do
+          expect(error.message(locale: :fr)).to eq 'Exemple'
+        end
+      end
+    end
+
+    context 'message is defined with i18n, but has no default translation' do
+      let(:error_class) do
+        Class.new(described_class) do
+          message 'Default message', i18n_key: 'foo'
+        end
+      end
+
+      it 'is equal to default message' do
+        expect(error_class.new.message(locale: :en)).to eq 'Default message'
+      end
+    end
+
+    context 'message is defined only with i18n, but has no default translation' do
+      let(:error_class) do
+        Class.new(described_class) do
+          message i18n_key: 'errors.foo'
+        end
+      end
+
+      it 'is equal to translation missing error text' do
+        expect(error_class.new.message(locale: :en)).to eq 'translation missing: en.errors.foo'
+      end
+    end
+
+    context 'message is defined with i18n that has interpolation' do
+      subject(:message) { error_class.new(variable: 'FOO', extra: 'bar').message(locale: :en) }
+
+      let(:error_class) do
+        Class.new(described_class) do
+          message 'Static default message', i18n_key: 'errors.variable'
+        end
+      end
+
+      it 'contains details in translation' do
+        is_expected.to eq 'Example with "FOO" variable'
       end
     end
 
@@ -242,16 +301,32 @@ module LunaPark
     describe '.message' do
       let(:error_class) do
         Class.new(described_class) do
-          message 'Text', i18n_key: 'errors.example_key'
+          message 'Static default message', i18n_key: 'errors.example_key'
         end
       end
 
-      it 'define default action' do
-        expect(error_class.default_message).to eq 'Text'
+      it 'defines default_message_block' do
+        expect(error_class.default_message_block).to be_a Proc
       end
 
-      it 'define 118n_key' do
+      it 'defines default_message_block that builds expected message' do
+        expect(error_class.default_message_block.call({})).to eq 'Static default message'
+      end
+
+      it 'defines 118n_key' do
         expect(error_class.i18n_key).to eq 'errors.example_key'
+      end
+
+      context 'when use default message block' do
+        let(:error_class) do
+          Class.new(described_class) do
+            message(i18n_key: 'errors.example_key') { |d| "Dynamic message #{d[:foo]}" }
+          end
+        end
+
+        it 'defines default_message_block that builds expected message' do
+          expect(error_class.default_message_block.call({ foo: 'Foo' })).to eq 'Dynamic message Foo'
+        end
       end
 
       context 'when use undefined action value' do
@@ -275,74 +350,46 @@ module LunaPark
       end
     end
 
-    describe '.default_message' do
-      subject(:default_message) { error_class.default_message }
+    describe '.default_message_block' do
+      subject(:default_message_block) { error_class.default_message_block }
 
       context 'when default message is not defined' do
         let(:error_class) { described_class }
 
-        it 'is eq error class name' do
-          is_expected.to eq error_class.name
+        it 'is nil' do
+          is_expected.to be_nil
         end
       end
 
       context 'when default message is defined' do
         let(:error_class) do
           Class.new(described_class) do
-            message 'Defined message'
+            message 'Static default message'
           end
         end
 
-        it 'is eq defined message' do
-          is_expected.to eq 'Defined message'
+        it 'is a Proc' do
+          is_expected.to be_a Proc
+        end
+
+        it 'can build defined message' do
+          expect(default_message_block.call({})).to eq 'Static default message'
         end
       end
 
-      context 'when i18n key is defined' do
+      context 'when default message is defined with block' do
         let(:error_class) do
           Class.new(described_class) do
-            message i18n_key: 'errors.example'
+            message { |d| "Message with detail '#{d[:detail]}'" }
           end
         end
 
-        it 'is eq error class name' do
-          is_expected.to eq error_class.name
-        end
-      end
-    end
-
-    describe '.translate' do
-      subject(:translate) { error_class.translate }
-
-      context 'when i18n key is not defined' do
-        let(:error_class) { described_class }
-
-        it { is_expected.to eq nil }
-      end
-
-      context 'when i18n_key is defined' do
-        let(:error_class) do
-          Class.new(described_class) do
-            message i18n_key: 'errors.example'
-          end
+        it 'is a Proc' do
+          is_expected.to be_a Proc
         end
 
-        it 'is eq translation of default locale' do
-          is_expected.to eq 'Example'
-        end
-      end
-
-      context 'when i18n_key is defined and selected locale' do
-        subject(:translate) { error_class.translate locale: :ru }
-
-        let(:error_class) do
-          Class.new(described_class) do
-            message i18n_key: 'errors.example'
-          end
-        end
-
-        it 'is eq  translation of selected locale' do
-          is_expected.to eq 'Пример'
+        it 'can build expected message' do
+          expect(default_message_block.call({ detail: 'DETAIL' })).to eq "Message with detail 'DETAIL'"
         end
       end
     end